+++ backends/backend-teavm/emu/com/badlogic/gdx/utils/SharedLibraryLoaderEmu.java
@@ -1,6 +1,7 @@
 package com.badlogic.gdx.utils;
 
 import com.badlogic.gdx.Gdx;
+import com.github.xpenatan.gdx.backends.teavm.assetloader.AssetInstance;
 import com.github.xpenatan.gdx.backends.teavm.assetloader.AssetLoaderListener;
 import com.github.xpenatan.gdx.backends.teavm.TeaApplication;
 import com.github.xpenatan.gdx.backends.teavm.gen.Emulate;
@@ -11,7 +12,7 @@ public class SharedLibraryLoaderEmu {
 
     public void load (String libraryName) {
         TeaApplication app = (TeaApplication)Gdx.app;
-        AssetLoader assetLoader = AssetLoader.getInstance();
+        AssetLoader assetLoader = AssetInstance.getLoaderInstance();
         assetLoader.loadScript(false, libraryName + ".js", new AssetLoaderListener<>() {
             @Override
             public void onSuccess(String url, String result) {